requirements.txt
yt-dlp
opencv-python
pytesseract
Pillow
